Add files via upload
Updated and upgraded a large amount of code. 
Added proper maze testing algorithm so that the program can insure that the maze can be solved. 
Made the maze navigation a separate function for easier referral and repeated use.
Added "revealed" attribute to rooms and restructured the map viewing functions to use that instead of "visited".
Added additional emoji and built a new system to help make it clear on what tile you are currently on.
Touched up the "end" command so that it requires two uses to quit.
Added additional clarifying documentation.
Fixed several small bugs that were due to misplaced brackets.
Tweaked the mapFunction to limit the width of the maze.
Improved help messages
Changed the map building so the exit was not initially revealed. 
Added hint command to reveal the exit if desired.